+++ test/specWindow-model-commandinput.js
@@ -2,7 +2,7 @@ var Application = require('spectron').Application;
 var assert = require('assert');
 var temp = require('temp').track();
 
-describe('systemWindow', function() {
+describe('specWindow-model-YCommandInput', function() {
   this.timeout(10000);
 
   before(function() {
@@ -24,7 +24,7 @@ describe('systemWindow', function() {
   beforeEach(function() {
     this.client = this.app.client;
     return this.client
-      .click('#show-system-window')
+      .click('#show-spec-window')
       .windowByIndex(1);
   });
 
@@ -32,16 +32,9 @@ describe('systemWindow', function() {
     return this.client.close();
   });
 
-  it('systemWindow load config', function() {
+  // TODO replace
+  it('YCommandInput', function() {
     return this.client
-      .getValue('#main-width').then(function(value) {
-        assert.ok(value);
-      })
-      .getValue('#main-height').then(function(value) {
-        assert.ok(value);
-      });
+      .click('#get-phont-list');
   });
-
-  // TODO save config
-  // TODO load config
 });
